% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{addDefault}
\alias{addDefault}
\title{Add default options if unset.}
\usage{
addDefault(options)
}
\arguments{
\item{options}{an options list}
}
\value{
an updated options list with the default records added for
unspecified fields
}
\description{
Add default options if unset.
}
\examples{
opts <- addDefault(list())

}
