% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{addKnownToFormulas}
\alias{addKnownToFormulas}
\title{Evaluate formulas in the environment of known params from the conditions}
\usage{
addKnownToFormulas(formulas, formulaIndexes, conditions)
}
\arguments{
\item{formulas}{a named list with unevaluated expressions
for expression levels
(e.g. describing amounts of "labelled", "total" etc. RNA)}

\item{formulaIndexes}{a list describing which formulas to use for
mean read number calculation for fractions defined by the names of list items}

\item{conditions}{a data.frame with the first column corresponding to
names in formulasIndexes (e.g. "total", "pull_down")}
}
\value{
a list with two items:
\itemize{
\item list of partially evaluated formulas
\item a vector of conditions
}
}
\description{
Evaluate formulas in the environment of known params from the conditions
}
