% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{amount}
\alias{amount}
\title{Create a formula object for the initial RNA level.}
\usage{
amount(x)
}
\arguments{
\item{x}{a character or a language object (an expression, a call
or a name)}
}
\value{
an expression for the initial RNA level
}
\description{
Create a formula object for the initial RNA level.
}
\examples{
x <- amount("mu")
degrade_(x, d_rate, time)
# mu_0 * exp(-d_rate * time)
}
