% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{amount_}
\alias{amount_}
\title{Create a formula object for the initial RNA level.}
\usage{
amount_(x)
}
\arguments{
\item{x}{an expression level}
}
\value{
an expression for the initial RNA level
}
\description{
A non-standard evaluation version of the \code{\link{amount}}
function.
}
