% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster_pca.R
\name{bhattacharyya_dist}
\alias{bhattacharyya_dist}
\title{Calculate Bhattacharyya Distance between two Multivariate Normal Distributions}
\usage{
bhattacharyya_dist(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1}{Numeric vector. Mean vector of the first cluster.}

\item{sigma1}{Matrix. Covariance matrix of the first cluster.}

\item{mu2}{Numeric vector. Mean vector of the second cluster.}

\item{sigma2}{Matrix. Covariance matrix of the second cluster.}
}
\value{
A single numeric value representing the distance.
}
\description{
Computes the Bhattacharyya distance, which measures the similarity of two probability
distributions. This implementation assumes multivariate normal distributions.
}
