% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{checkThresholds}
\alias{checkThresholds}
\title{Tests tolerance thersholds for correctness}
\usage{
checkThresholds(options)
}
\arguments{
\item{options}{an options list}
}
\description{
Throws an error in case wrong format or value are provided.
}
\keyword{internal}
