% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{codeFractions}
\alias{codeFractions}
\title{Generate fraction names}
\usage{
codeFractions(conditions, fractions)
}
\arguments{
\item{conditions}{a data.frame. The condition matrix.}

\item{fractions}{a formula, with the right side defining 
how to divide samples into different fractions, e.g. ~ time + conditions.}
}
\value{
a factor
}
\description{
Generate fraction names
}
