% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{ciGene}
\alias{ciGene}
\alias{ci}
\title{Estimate confidence intervals}
\usage{
ciGene(pd, fit, geneIndexes, parName, options, interval, confidence = 0.95)

ci(pd, fit, paramPath, options, interval, confidence = 0.95)
}
\arguments{
\item{geneIndexes}{a vector;  corresponds to the indexes of genes, for which the
confidence intervals must be computed}

\item{interval}{a vector of too numbers, define (min, max) allowed
parameter values}

\item{confidence}{confidence level for interval estimation}
}
\value{
\itemize{
\item \code{ciGene}:
a data.frame with two columns (left, right) confidence boundaries
in the order of geneIndexes;
\item \code{ci}: a vector of two numbers;
}
}
\description{
Estimate confidence intervals
}
