% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{constructFormulas}
\alias{constructFormulas}
\title{Evaluate formulas according to parameters, given in
the condition \code{data.frame}}
\usage{
constructFormulas(formulas, conditions)
}
\arguments{
\item{formulas}{list, generated by \code{\link{MeanFormulas}}}

\item{conditions}{a data.frame; the first column corresponds to
to \code{names(formulas)}, conditions given by the formulas.
Other columns, if exist, contain parameter values, which are sample specific.
The order of rows corresponds to the order of samples.}
}
\value{
list; formulas with substituted parameters according to the values
in the \code{conditions} data.frame
}
\description{
Evaluate formulas according to parameters, given in
the condition \code{data.frame}
}
