% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{contaminate}
\alias{contaminate}
\title{Generate a new formula for the mean read number}
\usage{
contaminate(formulas, target_condition, contaminant_condition, coef_name)
}
\arguments{
\item{formulas}{list, generated by \code{\link{MeanFormulas}}}

\item{target_condition}{character; a condition, 
for which a new formulas is to be generated}

\item{contaminant_condition}{character; a condition, which reads can contaminate
fraction corresponding to \code{target_condition}}

\item{coef_name}{character; the name of the cross-contamination rate}
}
\value{
unevaluated call, as from \code{\link{quote}} function
}
\description{
If the condition A can be contaminated with reads from the condition B,
its expected mean read number will be \eqn{(1-p)r_a + p r_b}, where
\eqn{p} is cross-contamination rate and \eqn{r_a,r_b} are
expected mean read numbers for conditions A and B, correspondingly.
}
\examples{
\dontrun{
forms <- MeanFormulas(A=r_a, B=r_b)
forms[["A"]] <- contaminate(forms, "A", "B", "p")
}
}
