% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{degrade}
\alias{degrade}
\title{Create a formula for RNA degradation}
\usage{
degrade(x, d, t)
}
\arguments{
\item{x}{initial concentration.}

\item{d}{a degradation rate.}

\item{t}{a longitude of the modelled period.}
}
\value{
an expression for the calculation of the RNA level after
degradation during time \verb{t}.
}
\description{
All the arguments must be characters or language objects 
(expression, call or name).
}
\examples{
x <- amount("mu_0")
mu <- amount_(mu_new)
d <- "degradation_rate"
t <- "t_labelling"
degrade(x, d, t)
# mu_0 * exp(-degradation_rate * t_labelling)

}
