% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{degrade_}
\alias{degrade_}
\title{Create a formula for RNA degradation}
\usage{
degrade_(x, d, t)
}
\arguments{
\item{x}{initial concentration.}

\item{d}{a degradation rate.}

\item{t}{a longitude of the modelled period.}
}
\value{
an expression for the calculation of the RNA level after
degradation during time \verb{t}.
}
\description{
Implements a non-standard evaluation version.
}
\examples{
x <- amount("a")
degrade_(x,b,c)
# a * exp(-b * c)

}
