% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{.eco_axis_relabel}
\alias{.eco_axis_relabel}
\title{Make alluvial stratum labels unique & readable per axis}
\usage{
.eco_axis_relabel(x, axis)
}
\description{
Make alluvial stratum labels unique & readable per axis
}
\keyword{internal}
