% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{.eco_choropleth}
\alias{.eco_choropleth}
\title{Internal: a discrete or continuous choropleth on an sf layer}
\usage{
.eco_choropleth(polys, col, discrete, title, sub, legend, border = "white")
}
\description{
Internal: a discrete or continuous choropleth on an sf layer
}
\keyword{internal}
