% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{.eco_rt_class_table}
\alias{.eco_rt_class_table}
\title{Parse the wide river_types column names "Region{i}RiverType{k}"}
\usage{
.eco_rt_class_table(rt)
}
\description{
Parse the wide river_types column names "Region{i}RiverType{k}"
}
\keyword{internal}
