% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{.eco_rt_hard}
\alias{.eco_rt_hard}
\title{Dominant (region, river-type) label per observation}
\usage{
.eco_rt_hard(rt)
}
\description{
Dominant (region, river-type) label per observation
}
\keyword{internal}
