% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{.region_axis_relabel}
\alias{.region_axis_relabel}
\title{Make alluvial stratum labels unique & readable per axis}
\usage{
.region_axis_relabel(x, axis)
}
\arguments{
\item{x}{Vector of labels to relabel (coerced to character).}

\item{axis}{One of \code{"region"}, \code{"local"} or \code{"mosaic"}; any
other value gets an abbreviated prefix.}
}
\value{
A character vector of prefixed, disambiguated labels.
}
\description{
Make alluvial stratum labels unique & readable per axis
}
\keyword{internal}
