% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{estimateCIFixed}
\alias{estimateCIFixed}
\title{Estimate CI when parameters of other genes are fixed}
\usage{
estimateCIFixed(pd, fit, geneIndexes, parName, options, confidence = 0.95,
  interval)
}
\arguments{
\item{pd}{a \link{PulseData} object}

\item{geneIndexes}{an integer vector; for which genes subset confidence
intervals are to be calculated}

\item{parName}{a character, e.g. "mu"}

\item{options}{an option list used for the \link{fitModel} call}

\item{confidence}{confidence level for interval estimation}

\item{interval}{}
}
\value{
a data.frame with two columns (left, right) confidence boundaries
in the order of geneIndexes
}
\description{
Estimate CI when parameters of other genes are fixed
}
