% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{findDeseqFactors}
\alias{findDeseqFactors}
\title{Calculate normalisation factors}
\usage{
findDeseqFactors(count_data, conditions, spikeins)
}
\arguments{
\item{count_data}{integer matrix, colnames correspond to samples 
(rownames in \code{conditions})}

\item{conditions}{factors to split samples for normalisation}

\item{spikeins}{integer of boolean vector for subseting genes 
(rows in \code{count_data}), which are to be used as spike-ins}
}
\value{
vector of double; normalisation factors in the same order as 
  columns in the \code{count_data}
}
\description{
Calculate normalisation factors
}

