% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{findDeseqFactorsForFractions}
\alias{findDeseqFactorsForFractions}
\title{Calculate normalisation factors}
\usage{
findDeseqFactorsForFractions(count_data, conditions)
}
\arguments{
\item{count_data}{integer matrix, colnames correspond to samples
(rownames in \code{conditions})}

\item{conditions}{factors to split samples for normalisation}
}
\value{
vector of double; normalisation factors in the same order as
columns in the \code{count_data}
}
\description{
Calculate normalisation factors
}
\keyword{internal}
