% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{findDeseqFactorsSingle}
\alias{findDeseqFactorsSingle}
\title{Calculate normalisation factors for columns in a matrix}
\usage{
findDeseqFactorsSingle(count_data)
}
\arguments{
\item{count_data}{a matrix; columns correspond to samples.}
}
\value{
a vector of doubles with normalisation factors ordered as columns in
the \code{count_data}
}
\description{
Calculate normalisation factors for columns in a matrix
}
\keyword{internal}
