% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitParams}
\alias{fitParams}
\alias{fitParamsSeparately}
\alias{fitNormFactors}
\alias{fitModel}
\title{Fit parameters given the initial values and the parameter names}
\usage{
fitParams(pd, par, namesToOptimise, options)

fitParamsSeparately(pd, par, knownGenePars, namesToOptimise, options,
  indexes = seq_len(dim(pd$counts)[1]))

fitNormFactors(pd, par, options)

fitModel(pulseData, par, options)
}
\arguments{
\item{pd}{the \code{\link{PulseData}} object}

\item{par}{the parameter named list}

\item{namesToOptimise}{a vector of names of parameters, which values
need be optimised}

\item{options}{a list with optimisation options. For details, see
\link{setTolerance}, \link{setFittingOptions}.}

\item{indexes}{indexes of genes to fit. By default includes all the genes.}

\item{pulseData}{a \link{PulseData} object}

\item{knownNames}{a vectors of names of the gene-specific parameters, which
are assumed to be fixed during optimisation.}

\item{par}{a list with an initial parameters values.
Gene-specific parameters must be set as vectors of the length equal to
the gene number. \link{initParameters} may simplify the process of initial
values randomisation.

The list must include \code{size} parameter for the negative binomial
distribution and a list with the normalisation factors, if
no spike-ins are used in the experiment.}

\item{options}{a list of options. For more details, see \link{setBoundaries},
\link{setTolerance}, \link{setFittingOptions}}
}
\value{
a list with fitted parameters

a list with fitted parameters

a list of normalisation factors

a list with the fitted parameters in the same form as
the initial guess \code{par} list
}
\description{
Fit parameters given the initial values and the parameter names

The same as \link{fitParams}, but performs optimisation for gene-specific
parameters only. Every set of parameters is fitted individually for
every gene.

Fit fraction normalisation coefficients

Fit the model by MLE
}
\examples{
\dontrun{
fitResult <- fitModel(pd, par)
}
}
