% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitDispersion}
\alias{fitDispersion}
\title{Fit the dispersion parameter}
\usage{
fitDispersion(pulseData, par, options)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{all needed parameters as a \code{list}}

\item{options}{\code{list} of options}
}
\value{
the MLE of the \code{size} for \code{\link{dnbinom}}
}
\description{
Fit the dispersion parameter
}
