% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitFractions}
\alias{fitFractions}
\title{Fit fraction normalisation coefficients}
\usage{
fitFractions(pulseData, par, options)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{all needed parameters as a \code{list}}

\item{options}{\code{list} of options}
}
\value{
vector of normalisation factors; \code{c(1,norm_factors)}
  corresponds to the fractions in \code{pulseData$fractions}
}
\description{
Fit fraction normalisation coefficients
}
