% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitGeneParameters}
\alias{fitGeneParameters}
\title{Fit gene-specific parameters}
\usage{
fitGeneParameters(pulseData, par, options)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{all needed parameters as a \code{list}}

\item{options}{\code{list} of options}
}
\value{
fitted parameters as a data.frame ordered as initial parameters
}
\description{
Fit gene-specific parameters
}
