% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitIndividualParameters}
\alias{fitIndividualParameters}
\title{Fit gene-specific parameters}
\usage{
fitIndividualParameters(pulseData, par, options)
}
\arguments{
\item{pulseData}{}

\item{par}{}

\item{options}{}
}
\value{
fitted parameters as a data.frame ordered as initial parameters
}
\description{
Fit gene-specific parameters
}

