% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit the model by MLE}
\usage{
fitModel(pulseData, par, options = list())
}
\arguments{
\item{pulseData}{}

\item{par}{}

\item{options}{}
}
\value{
a list `l` with fitted parameters `l$par` and
    formulas `l$formulas`
}
\description{
Fit the model by MLE
}
\examples{
\dontrun{
fitResult <- fitModel(pd, par)
}
}

