% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitNormFactors}
\alias{fitNormFactors}
\title{Fit fraction normalisation coefficients}
\usage{
fitNormFactors(pd, par, options)
}
\arguments{
\item{pd}{the \code{\link{PulseData}} object}

\item{par}{the parameter named list}

\item{options}{a list with optimisation options. For details, see
\link{setTolerance}, \link{setFittingOptions}.}
}
\value{
a list of normalisation factors
}
\description{
Fit fraction normalisation coefficients
}
