% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitParams}
\alias{fitParams}
\title{Fit parameters given the initial values and the parameter names}
\usage{
fitParams(pd, par, namesToOptimise, options)
}
\arguments{
\item{pd}{the \code{\link{PulseData}} object}

\item{par}{the parameter named list}

\item{namesToOptimise}{a vector of names of parameters, which values
need be optimised}

\item{options}{a list with optimisation options. For details, see
\link{setTolerance}, \link{setFittingOptions}.}
}
\value{
a list with fitted parameters
}
\description{
Fit parameters given the initial values and the parameter names
}
