% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitParamsSeparately}
\alias{fitParamsSeparately}
\title{Fit parameters with separate likelihood functions}
\usage{
fitParamsSeparately(pd, par, namesToOptimise, opts)
}
\arguments{
\item{pd}{the \code{\link{PulseData}} object}

\item{par}{the parameter named list}

\item{namesToOptimise}{a vector of names}

\item{opts}{a list with optimisation options}
}
\value{
a list with fitted parameters
}
\description{
Fit parameters with separate likelihood functions
}
