% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fittingOptions}
\alias{fittingOptions}
\title{Specify fitting options}
\usage{
fittingOptions(verbose = c("silent", "verbose"), cores = 1, options)
}
\arguments{
\item{verbose}{if "verbose" relative changes of parameters 
between two fitting iterations are printed}

\item{cores}{\code{integer}, natural number}

\item{options}{an option object}
}
\value{
an option object with modified specified parameters.
}
\description{
Specify fitting options
}
