% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{generateTestDataFrom}
\alias{generateTestDataFrom}
\title{Create a test count data}
\usage{
generateTestDataFrom(formulas, formulaIndexes, normFactors, par, conditions)
}
\arguments{
\item{formulas}{a list}

\item{formulaIndexes}{list of lists; defines indexes of formulas
used for calculation of the expected read number}

\item{normFactors}{list of vectors; normalisation factors, if
known}

\item{par}{a list of named parameters; gene-specific parameters
are vectors}

\item{conditions}{a condition data.frame}
}
\value{
matrix of counts with the order of columns as in conditions
}
\description{
Create a test count data
}
