% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{getCI}
\alias{getCI}
\title{Estimate confidence interval}
\usage{
getCI(pL, optimum, confidence, interval)
}
\arguments{
\item{pL}{a profile likelihood function}

\item{optimum}{an optimal value of the parameter (e.g. from fitting)}

\item{confidence}{confidence level for interval estimation}

\item{interval}{a vector of too numbers, define (min, max) allowed
parameter values}
}
\value{
a vector of two: c(min, max);
}
\description{
Estimate confidence interval
}
\details{
if the likelihood does not reach
the needed level (i.e. the CI exceeds the limitations in the \code{interval},
NA is returned. Hence, in case of both ends, the return value is c(NA, NA).
}
\keyword{internal}
