% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{getFractions}
\alias{getFractions}
\title{Get fraction labels for the samples}
\usage{
getFractions(pulseData)
}
\arguments{
\item{pulseData}{a \code{\link{PulseData}} object}
}
\value{
a vector of factors corresponding to the fraction 
   labels for samples, if fractions were specified in the \code{pulseData}
   argument. Otherwise, return \code{NULL}. 
   Sample labels are provided in the vector names.
}
\description{
Get fraction labels for the samples
}
\seealso{
\link{names}, \link{PulseData}
}
