% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{grow}
\alias{grow}
\title{Creates a formula which describe evolution of RNA concentration}
\usage{
grow(x, mu, d, t)
}
\arguments{
\item{x}{initial concentration.}

\item{mu}{new steady-state level.}

\item{d}{a degradation rate.}

\item{t}{a longitude of the modelled period.}
}
\value{
an expression for the calculation of the RNA level after time
\verb{t}.
}
\description{
All the arguments must be characters or language objects
(expression, call or name).
}
\examples{
x <- amount_(mu_0)
mu <- amount("mu_new")
d <- "degradation_rate"
t <- "t_labelling"
grow(x, mu, d, t)
# mu_new - (mu_new - mu_0) * exp(-degradation_rate * t_labelling)
}
