% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{grow_}
\alias{grow_}
\title{Creates a formula which describe evolution of RNA concentration}
\usage{
grow_(x, mu, d, t)
}
\arguments{
\item{x}{initial concentration.}

\item{mu}{new steady-state level.}

\item{d}{a degradation rate.}

\item{t}{a longitude of the modelled period.}
}
\value{
an expression for the calculation of the RNA level.
}
\description{
This implements the non-standard evaluation version of the
\code{\link{grow}} function.
}
\examples{
x <- amount("a")
grow_(x,b,c,d)
# b - (b - a) * exp(-c * d)

}
