% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{guessMeans}
\alias{guessMeans}
\title{Estimate initial guess for the mean expression level}
\usage{
guessMeans(pulseData, totalLabel = "total", fun = c("mean", "geomean",
  "median"))
}
\arguments{
\item{pulseData}{the \code{\link{PulseData}} object}

\item{totalLabel}{a character, the name of the factor level in the
condition matrix, which correspond to the total fraction ("total" by default).}

\item{fun}{a function used to estimate the expression level. 
Possible variants are mean, median and adjusted geometric mean (i.e.
$exp(mean(log(x + .5)))$.}
}
\value{
a vector of expression level estimations for every gene
}
\description{
Estimate initial guess for the mean expression level
}
\details{
Use this function to estimate mean read numbers
on the basis of the total fraction.
}
