% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{initParams}
\alias{initParams}
\title{Initialize first guess for the parameters}
\usage{
initParams(params, shared = NULL, fraction_factors = NULL, size = NULL)
}
\arguments{
\item{params}{a data.frame}

\item{shared}{a named list}

\item{fraction_factors}{a vector}

\item{size}{a double}
}
\value{
a list to provide to the function \link{\code{fitModel}}.
}
\description{
Initialize first guess for the parameters
}
