% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sample_means}
\alias{sample_means}
\alias{ll}
\alias{getNorms}
\alias{llnormFactors}
\alias{totalll}
\alias{predictExpression}
\alias{evaluateLikelihood}
\title{Get matrix for samples}
\usage{
sample_means(evaled_forms, norm_factors)

ll(par, namesToOptimise, pd, byOne = FALSE)

getNorms(pd, normFactors = NULL)

llnormFactors(par, pd)

totalll(par, pd, include = names(par))

predictExpression(par, pd)

evaluateLikelihood(par, pd)
}
\arguments{
\item{evaled_forms}{a numeric vector of formulas evaluated for every
condition}

\item{norm_factors}{a matrix n x k, where k is the number of samples}

\item{par}{a named list with parameter values (including the size parameter
for the negative binomial distribution, see \code{\link{dnbinom}})}

\item{namesToOptimise}{names of not fixed parameters}

\item{pd}{a \code{\link{PulseData}} object}

\item{byOne}{logical. If TRUE, the created function works on the level of
a single gene/isoform. In this case, every gene-specific parameter is
representd by a single scalar value. Otherwise, gene-specific parameters
are represented by vectors of the length equal to gene number.}

\item{normFactors}{a list of normalisation factors if inter-fraction
normalisation is used (i.e. no spike-ins are provided).
If NULL (default), only sequencing depth normalisation is used
on the basis of spike-ins counts. The structure of \code{normFactors} must be
same as of \code{pd$interSampleCoeffs}.}

\item{include}{character vector, names of the parameters in the par list,
which will be used in the created likelihood function}

\item{pd}{\code{\link{PulseData}} object}

\item{par}{estimated parameters from \link{fitModel}}

\item{pd}{a \link{PulseData} object.}
}
\value{
a vector of length equal to the sample number

a function with the folowing arguments:
\itemize{
\item x, a numeric vector of variable parameters, which is used to
calculate the likelihood function. The order of parameters is as in
unlist(par)
\item counts, a numeric matrix, or, if byOne is TRUE, a numeric vector with
read counts for every sample. If byOne is FALSE (default),
number of rows must be equal the number of genes.
\item fixedPars is a list with the rest of parameters which are needed for
likelihood calculation. By default, the ones provided in the \code{par}
argument to \code{ll} function are used.
}

The created function returns a logarithm of the likelihood value
calculated on the basis of the negative binomial distribution for the
provided counts and parameters.

a matrix of normalisation coefficients to use during calculation of
mean read number in samples. The row number equals number of
different formulas used in estimation of means
(i.e. the same as in \code{pd$rawFormulas}). The columns correspond to the samples
in the count matrix of the PulseData object \code{pd}.

a function with the following arguments:
\itemize{
\item x, a numeric vector which correspods to the records in
\code{pd$interSampleCoeffs}, but without the first element, i.e.
\code{unlist(pd$interSampleCoeffs)[-1]}
\item counts, a numeric matrix with read counts for every gene/isoform and
for every sample
}

The created function returns a logarithm of the likelihood function
calculated on the basis of the negative binomial distribution for the
provided counts, normalisation factors and  parameters.

a function with the following arguments:
\itemize{
\item x, a numeric vector which correspods to the records in
\code{par} argument. The first element of the normFactors, if
\code{par$normFactors} is not \code{NULL}, is excluded, because it is assumed
to be fixed to 1. Other parameters are the same as in \code{par}.
The structure of \code{x} is identical to \code{unlist(par[include])}. The first
element of \code{par$normFactors} is removed.
}

The created function returns a minus logarithm of the likelihood function
calculated on the basis of the negative binomial distribution for the
provided counts, normalisation factors and  parameters.

a named list:
\itemize{
\item predictions, a matrix of the same dimension as of the raw counts
\item llog, a matrix with logarithms of likelihood for the given raw counts.
}

a logarithm of the likelihood for given parameters and counts values.
}
\description{
Means are defined as a linear combination of the formulas, and
we use matrix representation of this relations via norm_factors matrix.
By multiplying evaluated formulas vector by norm_factors matrix,
we compute a weighted sum of the formulas.

Creates a likelihood function for given parameter names

Constructs a matrix of normalisation coefficients

The first element of the normalisation factors is assumed to be fixed
and equal 1. The rest of normalisation factors (i.e. without the first one)
are assumed as variable parameters for the likelihood function.

Create a likelihood function

Calculates mean read number estimations

Computes logarithm of the likelihood function
}
\examples{
\dontrun{
# Plot expected values vs the raw counts.
# Let res is the return of fitModel and pd is a PulseData object
pr <- predictExpression(pd, res)
plot(y = pr$predictions, x = pd$counts, xlab = "raw", ylab = "fitted")
}
}
\keyword{internal}
