% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{ll}
\alias{ll}
\title{Creates a likelihood function for given parameter names}
\usage{
ll(par, namesToOptimise, pulseData, byOne = FALSE)
}
\arguments{
\item{par}{a named list with parameter values (including the size parameter
for the negative binomial distribution, see \code{\link{dnbinom}})}

\item{namesToOptimise}{names of not fixed parameters}

\item{pulseData}{a \code{\link{PulseData}} object}

\item{byOne}{logical. If TRUE, the created function works on the level of
a single gene/isoform. In this case, every gene-specific parameter is
representd by a single scalar value. Otherwise, gene-specific parameters
are represented by vectors of the length equal to gene number.}
}
\value{
a function with the folowing arguments:
\itemize{
\item x, a numeric vector of variable parameters, which is used to
calculate the likelihood function. The order of parameters is as in
unlist(par)
\item counts, a numeric matrix, or, if byOne is TRUE, a numeric vector with
read counts for every sample. If byOne is FALSE (default),
number of rows must be equal the number of genes.
\item fixedPars is a list with the rest of parameters which are needed for
likelihood calculation. By default, the ones provided in the \code{par}
argument to \code{ll} function are used.
}

The created function returns a logarithm of the likelihood value
calculated on the basis of the negative binomial distribution for the
provided counts and parameters.
}
\description{
Creates a likelihood function for given parameter names
}
