% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{ll_gene}
\alias{ll_gene}
\title{Create a likelihood function for gene-specific parameters}
\usage{
ll_gene(pulseData, par)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{list; must have fields \code{shared_params}, \code{size},
\code{individual_params}, \code{fraction_factors}.}
}
\value{
a function(params, counts), which returns a  log likelihood
for a given vector of individual parameters, which are ordered as in 
\code{par$individual_params}.
}
\description{
The values of shared parameters, \code{size} from \code{\link{dnbinom}} and
normalisation factors are taken from \code{par}.
}
