% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{ll_shared_params}
\alias{ll_shared_params}
\title{Create a likelihood function for shared parameters}
\usage{
ll_shared_params(pulseData, par)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{list; must have fields \code{shared_params}, \code{size},
\code{individual_params}, \code{fraction_factors}.}
}
\value{
a function(params, counts), which returns a  log likelihood
for a given vector of shared parameters, which are ordered as in 
\code{par$shared_params}.
}
\description{
The values of gene-specific parameters, \code{size} from
 \code{\link{dnbinom}} and
normalisation factors are taken from \code{par}.
}
