% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{llnormFactors}
\alias{llnormFactors}
\title{Create a likelihood function for optimisation of normalisation factors}
\usage{
llnormFactors(par, pulseData)
}
\arguments{
\item{par}{a named list with parameter values (including the size parameter
for the negative binomial distribution, see \code{\link{dnbinom}})}

\item{pulseData}{a \code{\link{PulseData}} object}
}
\value{
a function with the following arguments:
\itemize{
\item x, a numeric vector which correspods to the records in
\code{pulseData$interSampleCoeffs}, but without the first element, i.e.
\code{unlist(pulseData$interSampleCoeffs)[-1]}
\item counts, a numeric matrix with read counts for every gene/isoform and
for every sample
}

The created function returns a logarithm of the likelihood function
calculated on the basis of the negative binomial distribution for the
provided counts, normalisation factors and  parameters.
}
\description{
The first element of the normalisation factors is assumed to be fixed
and equal 1. The rest of normalisation factors (i.e. without the first one)
are assumed as variable parameters for the likelihood function.
}
