% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{MeanFormulas}
\alias{MeanFormulas}
\title{Create list of formulas for expected read numbers}
\usage{
MeanFormulas(...)
}
\arguments{
\item{...}{list of named arguments; names represent condition labels,
the arguments values are unquoted expressions, which define formulas
for estimation of mean read number.}
}
\value{
list with keys corresponding to condition names and values as
  unevaluated call objects
}
\description{
Create list of formulas for expected read numbers
}
\examples{
\dontrun{
formulas <- MeanFormulas(total=mu, labelled=mu*d, unlabelled=mu*(1-d))
}
}
