% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{normalise.PulseData}
\alias{normalise.PulseData}
\title{Estimate normalisation factors for fraction with the same method as
in the DESeq2 package}
\usage{
\method{normalise}{PulseData}(pulseData)
}
\arguments{
\item{pulseData}{a PulseData object}
}
\value{
the same PulseData object with estimated normalisation factors
}
\description{
Estimate normalisation factors for fraction with the same method as
in the DESeq2 package
}

