% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{normalise}
\alias{normalise}
\title{Generic for normalisation of PulseData}
\usage{
normalise(pulseData)
}
\arguments{
\item{pulseData}{the object}
}
\description{
Generic for normalisation of PulseData
}

