% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{normaliseBoundaries}
\alias{normaliseBoundaries}
\title{Shape boundaries for the parameters in formulas}
\usage{
normaliseBoundaries(options, par, pd)
}
\arguments{
\item{options}{the options list}

\item{par}{the parameters list}

\item{pd}{the \code{\link{PulseData}} object}
}
\value{
an updated options list
}
\description{
If a single scalar value is provided, its boundaries are
assumed to be the same for all genes/isoforms, hence
a vector of gene number size will be returned.
}
