% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{normaliseNormFactorBoundaries}
\alias{normaliseNormFactorBoundaries}
\title{Shape boundaries for the normalisation factors
 
Create lower and upper boundaries with the same structure as
the list of normalisation coefficients `interSamplCoeffs` in
the \code{\link{PulseData}} object.}
\usage{
normaliseNormFactorBoundaries(options, pd)
}
\arguments{
\item{options}{the options list}

\item{pd}{the \code{\link{PulseData}} object}
}
\value{
an updated options list
}
\description{
The following cases for options${lb,ub}$normFactors are considered:
  - the structure is the same with the `interSampleCoeffs`
  - the length equals the number of unique conditions 
  - only a single scalar value is provided.
}
