% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{normaliseNoSpikeins}
\alias{normaliseNoSpikeins}
\title{Performs sequencing depth normalisation using the DESeq procedure.}
\usage{
normaliseNoSpikeins(pd, groups)
}
\arguments{
\item{pd}{a \code{\link{PulseData}} object}

\item{groups}{a vector for splitting objects to groups}
}
\value{
a vector with the coefficient for every sample
}
\description{
Performs sequencing depth normalisation using the DESeq procedure.
}
