% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{normaliseWithSpikeIns}
\alias{normaliseWithSpikeIns}
\title{Estimate DESeq normalisation factors using spike-ins counts}
\usage{
normaliseWithSpikeIns(pd, refGroup, spikeLists)
}
\arguments{
\item{pd}{a \code{\link{PulseData}} object}

\item{refGroup}{a character name of a sample groups to use as a reference.
The name must be present in the spikeLists names list}

\item{spikeLists}{a named list with the same structure as formulaIndexes}
}
\value{
a list of lists of the normalisation coefficients for every sample
}
\description{
Estimate DESeq normalisation factors using spike-ins counts
}
