% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{pl}
\alias{pl}
\title{Get the profile likelihood function (consider the parameters of other genes
as not fixed)}
\usage{
pl(paramPath, par, options, pd, namesToOptimise = names(options$lb))
}
\arguments{
\item{paramPath}{a list with names and indexes in order to locate the
parameter of the profile}

\item{par}{a result of the \link{fitModel} function}

\item{options}{an option list used for the \link{fitModel} call}

\item{pd}{a \link{PulseData} object}

\item{namesToOptimise}{which parameters are optimised (i.e. not fixed);
by default they are derived from the names of the boundaries}
}
\value{
a function with the calling convention as
\code{f(x = double(1)) --> double(1)},
which return the value of -log(likelihood(mu = x))
}
\description{
Get the profile likelihood function (consider the parameters of other genes
as not fixed)
}
\details{
\code{paramPath} is defined as the sequence of selections performed on
the parameter list (\code{par}). For example, list("mu", 1) will point to the
"mu" parameter of the first gene.
}
