% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{plFixed}
\alias{plFixed}
\title{Get the profile likelihood function (all other parameter fixed)}
\usage{
plFixed(parName, par, options, pd, geneIndex)
}
\arguments{
\item{parName}{a character, e.g. "mu"}

\item{par}{a result of the \link{`fitModel`} function}

\item{options}{an option list used for the \link{`fitModel`} call}

\item{pd}{a \link{`PulseData`} object}

\item{geneIndex}{an integer(1); a row index which corresponds to
the investigating gene}
}
\value{
a function with the calling convention as
\code{f(x = double(1)) --> double(1)},
which return the value of -log(likelihood(mu = x))
}
\description{
Get the profile likelihood function (all other parameter fixed)
}
