% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{plot_ecoregion_tuning}
\alias{plot_ecoregion_tuning}
\title{Combined eco-region tuning dashboard (surface + stability)}
\usage{
plot_ecoregion_tuning(eco)
}
\arguments{
\item{eco}{Result of \code{skater_con()}.}
}
\description{
Combined eco-region tuning dashboard (surface + stability)
}
