% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{plot_landscape_ksweep}
\alias{plot_landscape_ksweep}
\title{Joint scale (k-order) x cluster-number sweep heatmap}
\usage{
plot_landscape_ksweep(sweep_df)
}
\arguments{
\item{sweep_df}{Output of \code{sweep_k_order()}.}
}
\value{
A ggplot object.
}
\description{
Joint scale (k-order) x cluster-number sweep heatmap
}
