% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{plotPL}
\alias{plotPL}
\title{Plot the profile likeliihood}
\usage{
plotPL(pl, confidence = 0.95, ...)
}
\arguments{
\item{pl}{a result from the \link{profileGene} frunction}

\item{confidence}{a confidence level for the likelihood threshold line
(default .95)}

\item{...}{other arguments to the \code{\link[graphics]{plot}} function}
}
\value{
used for its side effect
}
\description{
Plot the profile likeliihood
}
