% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{profile}
\alias{profile}
\title{Profile}
\usage{
profile(paramPath, pd, par, options, interval,
  parName = as.character(paramPath[[1]]),
  namesToOptimise = names(options$lb), ...)
}
\arguments{
\item{paramPath}{a list with names and indexes in order to locate the
parameter of the profile}

\item{pd}{the \link{`PulseData`} object}

\item{par}{a result of the \link{`fitModel`} function}

\item{options}{the options list used for the \link{`fitModel`} function call}

\item{interval}{double(2) vector; left and right boundaries to calculate the
profile likelihood for the parameter given in \code{parName}}

\item{parName}{}

\item{namesToOptimise}{which parameters are optimised (i.e. not fixed);
by default they are derived from the names of the boundaries}

\item{...}{other parameters to pass to \link{`runPl`},
i.e. \code{logScale} (logical) and number of points \code{numPoints}}
}
\value{
a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.
}
\description{
Profile
}
