% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{profileOnlyGene}
\alias{profileOnlyGene}
\title{Estimate profile likelihood for gene parameters (all other fixed)}
\usage{
profileOnlyGene(pd, par, geneIndex, parName, options, interval, ...)
}
\arguments{
\item{pd}{the \link{`PulseData`} object}

\item{geneIndex}{integer(1); the row index in the count table
where the data for the given gene are located}

\item{parName}{a character;
the names of the gene-specific parameter (e.g. "mu")}

\item{options}{the options list used for the \link{`fitModel`} function call}

\item{interval}{double(2) vector; left and right boundaries to calculate the
profile likelihood for the parameter given in \code{parName}}

\item{...}{other parameters to pass to \link{`runPl`},
i.e. \code{logScale} (logical) and number of points \code{numPoints}}

\item{fit}{result output from the \link{`fitModel`} function}
}
\value{
a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.
}
\description{
Estimate profile likelihood for gene parameters (all other fixed)
}
