% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{profiles}
\alias{profiles}
\alias{plGene}
\alias{pl}
\title{Return a profile likelihood function for further use.}
\usage{
plGene(parName, geneIndex, par, pd, options)

pl(paramPath, par, pd, options, freeParams = names(options$lb))
}
\arguments{
\item{parName}{a character, e.g. "mu"}

\item{geneIndex}{an integer(1); a row index which corresponds to
the investigating gene}

\item{par}{a result of the \link{fitModel} function}

\item{pd}{a \link{PulseData} object}

\item{options}{an option list used for the \link{fitModel} call;
additional options can be specified:
\itemize{
\item jitter, a double (default: .1)
\item replicate, an integer number of repeating the optimisation from random
points
\item absolute, a logical (default: FALSE); if FALSE, the likelihood value at
the optimal point (i.e. \code{par}) is substracted from the returned value.
In this case, the value of the returned function at the point \code{par} is 0.
}}

\item{paramPath}{a list with names and indexes in order to locate the
parameter in the \code{par} argument (e.g. \code{list("mu", 1)} corresponds to
the "mu" parameter value for the first gene, i.e.
\code{\verb{par[["mu"]][[1]]}}}

\item{freeParams}{which parameters are optimised (i.e. not fixed);
by default they are derived from the names of the boundaries}
}
\value{
a function with the calling convention as
\code{f(x = double(1)) --> double(1))},
which return the value of -log(likelihood(mu = x))
}
\description{
A profile likelihood function returned by the \code{pl} considers parameters of
other genes. In contrast, the one from the \code{plGene} assumes
parameters of other genes,  shared parameters and
the normalisation factors to be fixed. \cr
The profile likelihood is estimated by numerical optimisation, and it can
be sensitive to the initial values. The optimisation is repeated
\code{options$replicate} times (default: 10), by adding a random values to the
the optimum  parameters at a scale specified in the options$jitter.
}
\details{
The randomisation of the parameter x is made as
$ x^* = (1 + a) * x$, where $a$ is a random value from the uniform
distribution (0,\code{options$jitter}).
}
