% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pulseRFractionData}
\alias{pulseRFractionData}
\title{An example of count data for an experiment without spike-ins}
\format{A list containing simulated data, model and parameter information.
\itemize{
\item \code{formulas} describe the model for mean read number
\item \code{counts} contain the simulated data
\item \code{conditions} is a data.frame describing sample time point and type
\item \code{fractions} is a character vector to divide samples into groups for
normalisation
\item \code{formulaIndexes} is a list of formula names (see \code{formulas}), which are
used in linear combination with coefficients in \code{normFactors} to calculate
mean read number
\item \code{normFactors} are coefficients which relate quantities of RNA between
different groups and can be used as normalisation coefficients for
calculation of mean read numbers using \code{formulas} and \code{formulaIndexes}
\item \code{par} is a list of model parameters
}}
\usage{
pulseRFractionData
}
\description{
An example of count data for an experiment without spike-ins
}
\details{
The data set contains simulation of an experiment which measures three
fractions, namely, coded as 'A_fraction', 'B_fraction' and 'C_fraction'.
There are three types of quantities, 'A', 'B' and 'C' for which a kinetic
model is defined, see \code{formulas} element in the data set:

(A = a * p, B =  a * b^time, C = a * (1 - b^time)),

where a, b are gene-specific parameters which are unknown, p is
a vector of known gene-specific parameters (e.g. probability of capture
depending on the uridine content etc.)

The data are generated for 3 replicates, 3 different time points and for
10 different genes, see elements \code{counts} and \code{conditions}.

The model considers possibility of cross-contamination with different types
of RNA, which is described by \code{formulaIndexes}. In this case,
the mean read number for a gene is a linear combination of the described
RNA types with weights defined in \code{normFactors}.

Due to different amounts of RNA in different conditions, the normalisation
factors are defined for groups of the samples, hence they are shared values
for a given sample group. The grouping is defined by \code{groups} element of the
data set.

The true parameter values, which are used for data simulation are in
\code{par} element. This also includes the size parameter for the negative
binomial distribution.
}
\keyword{datasets}
