% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pulseRSpikeinsData}
\alias{pulseRSpikeinsData}
\title{An example of count data for an experiment with spike-ins}
\format{A list containing simulated data, model and parameter information.
\itemize{
\item \code{formulas} describe the model for mean read number
\item \code{counts} contain the simulated data
\item \code{conditions} is a data.frame describing sample time point and type
\item \code{fractions} is a character vector to divide samples into groups for
normalisation
\item \code{formulaIndexes} is a list of formula names (see \code{formulas}), which are
used in linear combination with coefficients in \code{normFactors} to calculate
mean read number
\item \code{spikeins} is a list with two elements:
\itemize{
\item \code{refGroup} is the name of the sample which to use as a reference sample
\item \code{spikeLists} is a list of names of spike-ins which are to be used
for normalisation. It has the same structure as \code{formulaIndexes}
}
\item \code{allNormFactors} is a list of true coefficients which were used for
spike-ins and sample counts normalisations before simulation
\item \code{par} is a list of model parameters
}}
\usage{
pulseRSpikeinsData
}
\description{
An example of count data for an experiment with spike-ins
}
\keyword{datasets}
