% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{runPL}
\alias{runPL}
\title{Compute profile likelihood on the interval}
\usage{
runPL(pL, interval, logScale = FALSE, numPoints = 21)
}
\arguments{
\item{pL}{a function (x) --> double(1) returning the -log(likelihood).
For example, it can be a function returned by
\link{pl} or \link{plGene}.}

\item{interval}{double(2) vector; left and right boundaries to calculate the
profile likelihood for the parameter given in \code{parName}}

\item{logScale}{a logical (default: \code{FALSE}). Should points on the
\code{interval} be positioned at log scale?}

\item{numPoints}{the number of points to position at the \code{interval} for
profile likelihood calculations}
}
\value{
a data.frame; the first column consists of the parameter values,
the second ("logL") is for -log(likelihood) values.
}
\description{
Compute profile likelihood on the interval
}
