% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sample_means}
\alias{sample_means}
\title{Get matrix for samples}
\usage{
sample_means(evaled_forms, norm_factors)
}
\arguments{
\item{evaled_forms}{a numeric vector of formulas evaluated for every
condition}

\item{norm_factors}{a matrix n x k, where k is the number of samples}
}
\value{
a vector of length equal to the sample number
}
\description{
Means are defined as a linear combination of the formulas, and
we use matrix representation of this relations via norm_factors matrix.
}
\details{
For example, if there are pull-down and flow-through fractions,
which have 10% contamination by each other, and their mean read numbers
are 10 and 20 respectevily, the computation may look like
|.9 .1|
|10 20|x|     |
|.1 .9|
for one pull-down and one flow-through samples.
}
