% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{setBoundaries}
\alias{setBoundaries}
\title{Set optimization boundaries for the model parameters.}
\usage{
setBoundaries(b, normFactors = c(0.01, 10), options = .defaultParams)
}
\arguments{
\item{options}{an options object to use as a basis for a new parameter set}
}
\value{
an options object with the new parameter values
}
\description{
Set optimization boundaries for the model parameters.
}
\details{
If no options object is provided, the default value is used.
Boundaries are provided as a named list of vectors or lists with 
the length 2, see example.
}
\examples{
setBoundaries(params = list(a = c(1,2), b = c(10, 20)))

}
