% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{setFittingOptions}
\alias{setFittingOptions}
\title{Specify fitting options}
\usage{
setFittingOptions(verbose = c("silent", "verbose"), cores = 1, options)
}
\arguments{
\item{verbose}{if "verbose" relative changes of parameters 
between two fitting iterations are printed}

\item{cores}{\code{integer}, natural number}

\item{options}{an option object}
}
\value{
an option object with modified specified parameters.
}
\description{
Specify fitting options
}
