% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{setTolerance}
\alias{setTolerance}
\title{Set the stopping criteria in a form of the relative 
changes during fitting iterations.}
\usage{
setTolerance(params = NULL, shared = NULL, fraction_factors = NULL,
  size = NULL, options = .defaultParams)
}
\arguments{
\item{params}{double}

\item{shared}{double}

\item{fraction_factors}{double}

\item{size}{double}

\item{options}{an options object to use as a basis for a new parameter set}
}
\value{
an options object with the new parameter values
}
\description{
Set the stopping criteria in a form of the relative 
changes during fitting iterations.
}
\details{
If no options object is provided, the default value is used.
}
