% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{shrinkList}
\alias{shrinkList}
\title{Leave only one item per name}
\usage{
shrinkList(list)
}
\arguments{
\item{list}{a list to process}
}
\value{
a shorter list
}
\description{
If the same name is used for several items (e.g. "total"),
only first item is left in the result.
}
\examples{
l <- list(
  total   = 1,
  total   = 1,
  unlabel = c(4,5))
shrinkList(l)
}
