% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{stopIfNotInRanges}
\alias{stopIfNotInRanges}
\title{Validate list of parameters according to allowed value ranges.}
\usage{
stopIfNotInRanges(args, options)
}
\arguments{
\item{args}{a list of parametets}

\item{options}{an options object}
}
\description{
Validate list of parameters according to allowed value ranges.
}
