% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{theme_eco}
\alias{theme_eco}
\title{A clean shared theme for all ecotypology figures}
\usage{
theme_eco(base_size = 12, map = FALSE)
}
\arguments{
\item{base_size}{Base font size.}

\item{map}{Logical. If TRUE, strips axes/grid for choropleths.}
}
\description{
A clean shared theme for all ecotypology figures
}
