% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{validateOptions}
\alias{validateOptions}
\title{Check options for sanity.}
\usage{
validateOptions(options)
}
\arguments{
\item{options}{an options list}
}
\description{
Throws an error if incorrect values are supplied.
}
\keyword{internal}
