% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_components.R
\name{connect_components}
\alias{connect_components}
\title{Connect Disconnected Graph Components Using Spatial Proximity}
\usage{
connect_components(g, polygons)
}
\arguments{
\item{g}{An \code{igraph} object representing the graph with potentially
disconnected components.}

\item{polygons}{An \code{sf} or \code{sfc} object containing spatial polygons
corresponding to the nodes in the graph. The number of polygons must match
the number of vertices in \code{g}.}
}
\value{
An \code{igraph} object with all components connected. The returned
graph will have a single connected component.
}
\description{
Iteratively connects disconnected components of a graph by adding edges
between spatially closest nodes from different components. The function uses
polygon centroids to compute spatial distances and applies a minimum spanning
tree approach to ensure efficient connectivity.
}
\details{
The function works by:
\enumerate{
\item Computing centroids of the input polygons
\item Calculating the pairwise Euclidean distance matrix between all centroids
\item Identifying disconnected components in the graph
\item For each iteration while multiple components exist:
\itemize{
\item Computing minimum distances between all pairs of components
\item Creating a component-level graph weighted by these distances
\item Finding the minimum spanning tree (MST) of the component graph
\item Adding an edge between the two nodes that represent the shortest
connection between components
}
}

The algorithm continues until all components are connected into a single
connected component.
}
\examples{
\dontrun{
library(igraph)
library(sf)

# Create a simple graph with two components
g <- make_empty_graph(n = 6) |>
  add_edges(c(1, 2, 2, 3)) |>
  add_edges(c(4, 5, 5, 6))

# Create corresponding spatial polygons
polygons <- st_as_sf(data.frame(
  id = 1:6,
  geometry = st_sfc(
    st_point(c(0, 0)), st_point(c(1, 0)), st_point(c(2, 0)),
    st_point(c(10, 0)), st_point(c(11, 0)), st_point(c(12, 0))
  ) |> st_buffer(0.5)
))

# Connect the components
g_connected <- connect_components(g, polygons)
components(g_connected)$no  # Should be 1
}

}
\seealso{
\code{\link[igraph]{components}}, \code{\link[igraph]{mst}}
}
