% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{estimateCIFixed}
\alias{estimateCIFixed}
\title{Estimate CI when parameters of other genes are fixed}
\usage{
estimateCIFixed(pd, fit, geneIndexes, parName, options, confidence = 0.95,
  interval)
}
\arguments{
\item{pd}{}

\item{fit}{}

\item{geneIndexes}{}

\item{parName}{}

\item{options}{}

\item{confidence}{}

\item{interval}{}
}
\description{
Estimate CI when parameters of other genes are fixed
}
