% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{evaluateLikelihood}
\alias{evaluateLikelihood}
\title{Computes logarithm of the likelihood function}
\usage{
evaluateLikelihood(pulseData, par)
}
\arguments{
\item{par}{a named list with parameter values (including the size parameter
for the negative binomial distribution, see \code{\link{dnbinom}})}
}
\value{
a logarithm of the likelihood for given parameters and counts values.
}
\description{
Computes logarithm of the likelihood function
}
