% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit the model by MLE}
\usage{
fitModel(pulseData, par, options)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{initial guess for parameters as a \code{list}.}

\item{options}{\code{list} of options}
}
\value{
a list with the fitted parameters in the same form as
the initial guess \code{par}
}
\description{
Fit the model by MLE
}
\examples{
\dontrun{
fitResult <- fitModel(pd, par)
}
}
