% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit the model by MLE}
\usage{
fitModel(pulseData, par, options = .defaultParams)
}
\arguments{
\item{pulseData}{PulseData object}

\item{par}{initial guess for parameters as a \code{list}.}

\item{options}{\code{list} of options}
}
\value{
a list `l` with fitted parameters `l$par` and
    formulas `l$formulas`
}
\description{
Fit the model by MLE
}
\details{
In the initial guess \code{par},
   the following list items need to be provided:
   \itemize{
   \item{params: }{list of gene-specific parameters}
   \item{shared: }{list of shared parameters}
   \item{fraction_factors (if relevant): }{a vector of fraction factors
   in the order of \code{levels(pulseDatad$fractions)}} 
   \item{size (optional): }{an initial guess for the size parameter of the
   negative binomial distribution, see \code{\link{dnbinom}}}
   }
}
\examples{
\dontrun{
fitResult <- fitModel(pd, par)
}
}
