% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit the model by MLE}
\usage{
fitModel(pulseData, par, options)
}
\arguments{
\item{pulseData}{a \link{PulseData} object}

\item{par}{a list with an initial parameters values.
Gene-specific parameters must be set as vectors of the length equal to
the gene number. \link{initParameters} may simplify the process of initial
values randomisation.

The list must include \code{size} parameter for the negative binomial
distribution and a list with the normalisation factors, if
no spike-ins are used in the experiment.}

\item{options}{a list of options. For more details, see \link{setBoundaries},
\link{setTolerance}, \link{setFittingOptions}}
}
\value{
a list with the fitted parameters in the same form as
the initial guess \code{par} list
}
\description{
Fit the model by MLE
}
\examples{
\dontrun{
fitResult <- fitModel(pd, par)
}
}
