% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitParamsSeparately}
\alias{fitParamsSeparately}
\title{Fit parameters with separate likelihood functions}
\usage{
fitParamsSeparately(pd, par, knownNames, namesToOptimise, options,
  indexes = seq_len(dim(pd$counts)[1]))
}
\arguments{
\item{pd}{the \code{\link{PulseData}} object}

\item{par}{the parameter named list}

\item{knownNames}{a vectors of names of the gene-specific parameters, which
are assumed to be fixed during optimisation.}

\item{namesToOptimise}{a vector of names of parameters, which values
need be optimised}

\item{options}{a list with optimisation options. For details, see
\link{setTolerance}, \link{setFittingOptions}.}

\item{indexes}{indexes of genes to fit. By default includes all the genes.}
}
\value{
a list with fitted parameters
}
\description{
The same as \link{fitParams}, but performs optimisation for gene-specific
parameters only. Every set of parameters is fitted individually for
every gene.
}
