% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{generateTestDataFrom}
\alias{generateTestDataFrom}
\title{Create a test count data}
\usage{
generateTestDataFrom(formulas, par, conditions)
}
\arguments{
\item{formulas}{a list}

\item{par}{a list with individual_params(must have), size (must have) 
and shared_params (optional).}

\item{conditions}{a condition data.frame}
}
\value{
matrix of counts with the order of columns as in conditions
}
\description{
Create a test count data
}

