% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{generateTestDataFrom}
\alias{generateTestDataFrom}
\title{Create a test count data}
\usage{
generateTestDataFrom(formulas, par, conditions, fractions = NULL)
}
\arguments{
\item{formulas}{a list}

\item{par}{a list with individual_params(must have), size (must have) 
and shared_params (optional). If \code{fractions} is defined,
\code{par$fraction_factors} must be not \code{NULL}}

\item{conditions}{a condition data.frame}

\item{fractions}{a factor for splitting the samples into subsets with 
different normalisation coefficients}
}
\value{
matrix of counts with the order of columns as in conditions
}
\description{
Create a test count data
}

