% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{growFrom0}
\alias{growFrom0}
\title{Creates a formula which describe evolution of RNA concentration
if the initial amount is 0.}
\usage{
growFrom0(mu, d, t)
}
\arguments{
\item{mu}{new steady-state level.}

\item{d}{a degradation rate.}

\item{t}{a longitude of the modelled period.}
}
\value{
an expression to calculate the RNA level.
}
\description{
All the arguments must be characters or language objects 
(expression, call or name).
}
\examples{
growFrom0("b","c","d")
# b - b * exp(-c * d)

}
