% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{initParameters}
\alias{initParameters}
\title{Initialize first guess for the parameters}
\usage{
initParameters(par, geneParams, pulseData, options)
}
\arguments{
\item{par}{a list with parameter values}

\item{pulseData}{a \code{\link{PulseData}} object}

\item{options}{an options object}
}
\value{
a list to provide to the function \code{\link{fitModel}}.
}
\description{
Initialize first guess for the parameters
}
