% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{initParameters}
\alias{initParameters}
\title{Initialize first guess for the parameters}
\usage{
initParameters(par, geneParams, pulseData, options)
}
\arguments{
\item{par}{a list with parameter values}

\item{geneParams}{a vector of names of the gene-specific parameters.
For this parameters, the output length is equal the number of genes, i.e.
an individual parameter value is generated for every gene.}

\item{pulseData}{a \code{\link{PulseData}} object}

\item{options}{an options object}
}
\value{
a list to provide to the function \code{\link{fitModel}}.
}
\description{
If the normalisation factors are provided, i.e. \code{par$normFactors} is
not NULL, they are also randomly generated with the resulting structure
equal to the \code{pulseData$interSampleCoeffs}. Please make sure, that
the upper and lower boundaries are correctly named and structured.
}
