% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{initParams}
\alias{initParams}
\title{Initialize first guess for the parameters}
\usage{
initParams(options, params, shared = NULL, fraction_factors = NULL)
}
\arguments{
\item{options}{an options object}

\item{params}{a data.frame}

\item{shared}{a named list}

\item{fraction_factors}{a vector}
}
\value{
a list to provide to the function \code{\link{fitModel}}.
}
\description{
Initialize first guess for the parameters
}
