% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{initParams}
\alias{initParams}
\title{Initialize first guess for the parameters}
\usage{
initParams(pulseData, options, params = NULL, shared = NULL,
  fraction_factors = NULL, size = NULL)
}
\arguments{
\item{pulseData}{a \code{\link{PulseData}} object}

\item{options}{an options object}

\item{params}{a data.frame}

\item{shared}{a named list}

\item{fraction_factors}{a vector}

\item{size}{a double; size parameter for the \code{\link{dnbinom}}}
}
\value{
a list to provide to the function \code{\link{fitModel}}.
}
\description{
Initialize first guess for the parameters
}
