% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{predictExpression}
\alias{predictExpression}
\title{Calculates mean read number estimations}
\usage{
predictExpression(par, pd)
}
\arguments{
\item{par}{estimated parameters from \link{fitModel}}

\item{pd}{a \link{PulseData} object.}
}
\value{
a named list:
\itemize{
\item predictions, a matrix of the same dimension as of the raw counts
\item llog, a matrix with logarithms of likelihood for the given raw counts.
}
}
\description{
Calculates mean read number estimations
}
\examples{
\dontrun{
# Plot expected values vs the raw counts.
# Let res is the return of fitModel and pd is a PulseData object
pr <- predictExpression(pd, res)
plot(y = pr$predictions, x = pd$counts, xlab = "raw", ylab = "fitted")
}
}
