% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{predictExpression}
\alias{predictExpression}
\title{Calculate expected read numbers for the raw data}
\usage{
predictExpression(pulseData, par)
}
\arguments{
\item{pulseData}{a \link{PulseData} object.}

\item{par}{estimated parameters from \link{fitModel}}
}
\value{
a list(preditions=,llog=), where
  predictions is a matrix of the same dimension 
  as the raw counts in pulseData$count_data;
  llog is a matrix with logarithms of likelihood for the given raw counts.
}
\description{
Calculate expected read numbers for the raw data
}
