% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{predictExpression}
\alias{predictExpression}
\title{Calculates mean read number estimations}
\usage{
predictExpression(pulseData, par)
}
\arguments{
\item{pulseData}{a \link{PulseData} object.}

\item{par}{estimated parameters from \link{fitModel}}
}
\value{
a named list:
  - preditions, a matrix of the same dimension as of the raw counts 
  - llog, a matrix with logarithms of likelihood for the given raw counts.
}
\description{
Calculates mean read number estimations
}
