% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{profileGene}
\alias{profileGene}
\alias{profile}
\alias{plotPL}
\title{Estimate profile likelihood for gene parameters (all other fixed)}
\usage{
profileGene(parName, geneIndex, pd, par, interval, options, ...)

profile(paramPath, pd, par, options, interval,
  namesToOptimise = names(options$lb), ...)

plotPL(pl, confidence = 0.95, ...)
}
\arguments{
\item{parName}{a character;
the names of the gene-specific parameter (e.g. "mu")}

\item{geneIndex}{integer(1); the row index in the count table
where the data for the given gene are located}

\item{pd}{the \link{PulseData} object}

\item{par}{a result of the \link{fitModel} function}

\item{interval}{double(2) vector; left and right boundaries to calculate the
profile likelihood for the parameter given in \code{parName}}

\item{options}{the options list used for the \link{fitModel} function call}

\item{...}{other parameters to pass to \link{runPL},
i.e. \code{logScale} (logical) and number of points \code{numPoints}}

\item{paramPath}{a list with names and indexes in order to locate the
parameter in the \code{par} argument (e.g. \code{list("mu", 1)} corresponds to
the "mu" parameter value for the first gene, i.e.
\code{\verb{par[["mu"]][[1]]}}}

\item{pl}{a result from the \link{profileGene} frunction}

\item{confidence}{a confidence level for the likelihood threshold line
(default .95)}

\item{fit}{result output from the \link{fitModel} function}

\item{parName}{}
}
\value{
a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.

a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.

used for its side effect
}
\description{
Estimate profile likelihood for gene parameters (all other fixed)

Profile

Plot the profile likeliihood
}
