% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{profileOnlyGene}
\alias{profileOnlyGene}
\alias{profile}
\alias{runPL}
\alias{plFixed}
\alias{pl}
\alias{plotPL}
\alias{ciGene}
\alias{getCI}
\title{Estimate profile likelihood for gene parameters (all other fixed)}
\usage{
profileOnlyGene(pd, par, geneIndex, parName, options, interval, ...)

profile(paramPath, pd, par, options, interval,
  parName = as.character(paramPath[[1]]),
  namesToOptimise = names(options$lb), ...)

runPL(pL, interval, logScale = FALSE, numPoints = 21)

plFixed(parName, par, options, pd, geneIndex)

pl(paramPath, par, options, pd, namesToOptimise = names(options$lb))

plotPL(pl, confidence = 0.95, ...)

ciGene(pd, fit, geneIndexes, parName, options, confidence = 0.95, interval)

getCI(pL, optimum, confidence, interval)
}
\arguments{
\item{pd}{the \link{PulseData} object}

\item{par}{a result of the \link{fitModel} function}

\item{geneIndex}{integer(1); the row index in the count table
where the data for the given gene are located}

\item{parName}{a character;
the names of the gene-specific parameter (e.g. "mu")}

\item{options}{the options list used for the \link{fitModel} function call}

\item{interval}{double(2) vector; left and right boundaries to calculate the
profile likelihood for the parameter given in \code{parName}}

\item{...}{other parameters to pass to \link{runPL},
i.e. \code{logScale} (logical) and number of points \code{numPoints}}

\item{paramPath}{a list with names and indexes in order to locate the
parameter of the profile}

\item{namesToOptimise}{which parameters are optimised (i.e. not fixed);
by default they are derived from the names of the boundaries}

\item{pL}{a function (x) --> double(1) returning the -log(likelihood).
For example, it can be a function returned by
\link{`pl`} or \link{`plFixed`}.}

\item{logScale}{a logical (default: \code{FALSE}). Should points on the
\code{interval} be positioned at log scale?}

\item{numPoints}{the number of points to position at the \code{interval} for
profile likelihood calculations}

\item{pl}{a result from the \link{profileOnlyGene} frunction}

\item{confidence}{a confidence level for the likelihood threshold line
(default .95)}

\item{fit}{result output from the \link{fitModel} function}

\item{geneIndexes}{an integer vector; for which genes subset confidence
intervals are to be calculated}

\item{optimum}{an optimal value of the parameter (e.g. from fitting)}

\item{parName}{}

\item{parName}{a character, e.g. "mu"}

\item{options}{an option list used for the \link{fitModel} call}

\item{pd}{a \link{PulseData} object}

\item{geneIndex}{an integer(1); a row index which corresponds to
the investigating gene}

\item{par}{a result of the \link{fitModel} function}

\item{options}{an option list used for the \link{fitModel} call}

\item{pd}{a \link{PulseData} object}

\item{interval}{}

\item{pL}{a profile likelihood function}

\item{confidence}{confidence level for interval estimation}

\item{interval}{a vector of too numbers, define (min, max) allowed
parameter values}
}
\value{
a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.

a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.

a data.frame; the first column consists of the parameter values,
the second ("logL") is for -log(likelihood) values.

a function with the calling convention as
\code{f(x = double(1)) --> double(1)},
which return the value of -log(likelihood(mu = x))

a function with the calling convention as
\code{f(x = double(1)) --> double(1)},
which return the value of -log(likelihood(mu = x))

used for its side effect

a data.frame with two columns (left, right) confidence boundaries
in the order of geneIndexes

a vector of two: c(min, max);
}
\description{
Estimate profile likelihood for gene parameters (all other fixed)

Profile

Compute profile likelihood on the interval

Get the profile likelihood function (all other parameter fixed)

Get the profile likelihood function (consider the parameters of other genes
as not fixed)

Plot the profile likeliihood

Estimate CI when parameters of other genes are fixed

Estimate confidence interval
}
\details{
\code{paramPath} is defined as the sequence of selections performed on
the parameter list (\code{par}). For example, list("mu", 1) will point to the
"mu" parameter of the first gene.

if the likelihood does not reach
the needed level (i.e. the CI exceeds the limitations in the \code{interval},
NA is returned. Hence, in case of both ends, the return value is c(NA, NA).
}
\keyword{internal}
