% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{profileGene}
\alias{profileGene}
\alias{profile}
\title{Estimate profile likelihood for gene parameters (all other fixed)}
\usage{
profileGene(parName, geneIndex, pd, par, interval, options, ...)

profile(paramPath, pd, par, options, interval, namesToOptimise = names(par),
  ...)
}
\arguments{
\item{parName}{a character;
the names of the gene-specific parameter (e.g. "mu")}

\item{geneIndex}{integer(1); the row index in the count table
where the data for the given gene are located}

\item{pd}{the \link{PulseData} object}

\item{par}{result output from the \link{fitModel} function}

\item{interval}{a vector of two numbers defining the interval of
parameter profiling}

\item{options}{the options list used for the \link{fitModel} function call}

\item{...}{other parameters to pass to \link{runPL},
i.e. \code{logScale} (logical) and number of points \code{numPoints}}

\item{paramPath}{a list with names and indexes in order to locate the
parameter in the \code{par} argument (e.g. \code{list("mu", 1)} corresponds to
the "mu" parameter value for the first gene, i.e.
\verb{par[["mu"]][[1]]}.}

\item{namesToOptimise}{a character vector including the free parameters
(i.e. they wil be treated as unfixed)}
}
\value{
a data.frame; the column \code{logl} corresponds to the -log(likelihood)
function values. the other represents the profiled parameter values.
}
\description{
Estimate profile likelihood for gene parameters (all other fixed)
}
