% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseData.R
\name{PulseData}
\alias{PulseData}
\title{Create an object for pulse-change count data}
\usage{
PulseData(count_data, conditions, formulas, spikeins = NULL,
  fractions = NULL)
}
\arguments{
\item{count_data}{a matrix; column names correspond to sample names}

\item{conditions}{a data.frame;
the first column corresponds to the conditions given in \code{formulas}.
May also include column "fraction" and columns named as parameters,
used in the formula definitions, e.g. "time".}

\item{formulas}{a list, created by \code{\link{MeanFormulas}}}

\item{spikeins}{a vector of charecters or indexes, optional;
defines which genes to use as a reference for normalisation}

\item{fractions}{a formula, e.g. ~ condition + time (if spike-ins are
not provided).
the names used in the \code{fractions} defines different fractions,
which should have distinct coefficients for mean expression fitting.}
}
\value{
an object of class "PulseData"
}
\description{
Create an object for pulse-change count data
}

