% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pulseRFractionData}
\alias{pulseRFractionData}
\title{An example of count data for an experiment without spike-ins}
\format{A list containing simulated data, model and parameter information.
\itemize{
\item \code{formulas} describe the model for mean read number
\item \code{counts} contain the simulated data
\item \code{conditions} is a data.frame describing sample time point and type
\item \code{fractions} is a character vector to divide samples into groups for
normalisation
\item \code{formulaIndexes} is a list of formula names (see \code{formulas}), which are
used in linear combination with coefficients in \code{normFactors} to calculate
mean read number
\item \code{normFactors} are coefficients which relate quantities of RNA between
different groups and can be used as normalisation coefficients for
calculation of mean read numbers using \code{formulas} and \code{formulaIndexes}
\item \code{par} is a list of model parameters
}}
\usage{
pulseRFractionData
}
\description{
An example of count data for an experiment without spike-ins
}
\keyword{datasets}
