% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sample_means}
\alias{sample_means}
\title{Get matrix for samples}
\usage{
sample_means(evaled_forms, norm_factors)
}
\arguments{
\item{evaled_forms}{a numeric vector of formulas evaluated for every
condition}

\item{norm_factors}{a matrix n x k, where k is the number of samples}
}
\value{
a vector of length equal to the sample number
}
\description{
Means are defined as a linear combination of the formulas, and
we use matrix representation of this relations via norm_factors matrix.
By multiplying evaluated formulas vector by norm_factors matrix,
we compute a weighted sum of the formulas.
}
