% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{setBoundaries}
\alias{setBoundaries}
\title{Set optimization boundaries for the model parameters.}
\usage{
setBoundaries(params = NULL, shared = NULL, fraction_factors = NULL,
  size = NULL, options = .defaultParams)
}
\arguments{
\item{params}{a list of gene-specific parameter boundaries}

\item{shared}{a list of shared parameters boundaries}

\item{fraction_factors}{the  lower and the upper boundaries 
for the fraction factors}

\item{size}{the lower and the upper boundaries for the size parameter for
\code{\link{dnbinom}}.}

\item{options}{an options object to use as a basis for a new parameter set}
}
\value{
an options object with the new parameter values
}
\description{
Set optimization boundaries for the model parameters.
}
\details{
If no options object is provided, the default value is used.
Boundaries are provided as a named list of vectors or lists with 
the length 2, see example.
}
\examples{
setBoundaries(params = list(a = c(1,2), b = c(10, 20)))

}
