% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{setBoundaries}
\alias{setBoundaries}
\title{Set optimization boundaries for the model parameters.}
\usage{
setBoundaries(params = NULL, shared = NULL, fraction_factors = NULL,
  size = NULL, options = .defaultParams)
}
\arguments{
\item{params}{boundaries for gene-specific parameters}

\item{shared}{boundaries for shared parameters}

\item{fraction_factors}{boundaries for fraction factors if relevant}

\item{size}{boundaries for the size parameter of the negative binomial
ditribution}

\item{options}{an options object to use as a basis for a new parameter set}
}
\value{
an options object with the new parameter values
}
\description{
Set optimization boundaries for the model parameters.
}
\details{
If no options object is provided, the default value is used.
}
